WIP travis-ci, try 6

+++ tools/build_utils.py
@@ -195,7 +195,7 @@ def build_and_run_pytest(redo_cmake=False):
         proj_root = bytes(proj_root, 'ascii')
     args = [x for x in sys.argv[1:] if x.endswith('.py') and
             os.path.basename(x).startswith('test')]
-    ncpu = multiprocessing.cpu_count() / 2
+    ncpu = int(multiprocessing.cpu_count() / 2)
     if not args:
         args = ['.', '--ignore', 'build', '-n%s' % (ncpu)]
     else:  # running one file, don't scan
